Direct access control (defined WPINC) added.

+++ uninstall.php
@@ -25,6 +25,11 @@
  * @package    Speed_Booster_Pack
  */
 
+// If this file is called directly, abort.
+if ( ! defined( 'WPINC' ) ) {
+	die;
+}
+
 // If uninstall not called from WordPress, then exit.
 if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	delete_option( 'sbp_options' );
